requirements.txt
pytest
pytest-qt
PyHamcrest
pyhamcrest-qt
pyside2
mock
